Refactor animation into useAnimatedFooterStyle
